+++ RELEASENOTES.md
@@ -74,6 +74,8 @@
 *   Cast extension:
     *   Add support for `setVolume()`, and `getVolume()`
         ([#2279](https://github.com/androidx/media/pull/2279)).
+    *   Prevent CastPlayer from entering STATE_BUFFERING while the timeline is
+        empty.
 *   Test Utilities:
     *   Add `advance(player).untilPositionAtLeast` and `untilMediaItemIndex` to
         `TestPlayerRunHelper` in order to advance the player until a specified